refactoring tool prompting to avoid circular imports
