+++ _scripts/evaluate_doc_retrieval.py
@@ -1,7 +1,6 @@
 import logging
 import math
 from typing import List, NamedTuple, Union
-import argparse
 
 import json
 from langchain_core.tracers.context import tracing_v2_enabled
