+++ _scripts/evaluate_tools.py
@@ -1,8 +1,7 @@
-import argparse
-from datetime import datetime
+import ast
 import logging
 from typing import Tuple, Dict
-import ast
+
 import pandas as pd
 from langchain_core.tracers.context import tracing_v2_enabled
 from langchain_core.tracers.langchain import LangChainTracer
@@ -31,7 +30,7 @@ def evaluate_output(df):
     return df
 
 
-def get_actual_results(langchain_tracer: LangChainTracer, name: str) -> Tuple[Dict, Dict]:
+def get_actual_results(langchain_tracer: LangChainTracer, name: str) -> Tuple[str, str]:
     input_result = get_nodes(
         root_node=langchain_tracer.latest_run,
         node_name=name,
@@ -60,13 +59,13 @@ if __name__ == "__main__":
 
     if args.eval_path != "None":
         # getting eval df
-        eval_df = pd.read_csv(args.eval_path, sep=";")
+        eval_df = pd.read_csv(args.eval_path, sep=";").iloc[:2, :]
 
         # running each scenario
-        input_actual, output_actual = list(), list()
+        input_actual, output_actual, responses = list(), list(), list()
         for i, row in eval_df.iterrows():
             with tracing_v2_enabled() as langchain_tracer:
-                model.invoke({
+                response = model.invoke({
                     "chat_history": [],
                     "question": str(row["query"])
                 })
@@ -74,17 +73,17 @@ if __name__ == "__main__":
             input_result, output_result = get_actual_results(langchain_tracer, str(row["tool_used"]))
             input_actual.append(input_result)
             output_actual.append(output_result)
+            responses.append(response)
+
+
 
         # adding data to df
+        eval_df["full_chain_output"] = responses
+        eval_df["SQL_Query"] = [i.split("Output :")[0] for i in output_actual]
         eval_df["actual_arguments"] = input_actual
         eval_df["actual_output"] = output_actual
         eval_df["expected_arguments"] = eval_df["expected_arguments"].apply(ast.literal_eval)
         eval_df["actual_arguments"] = eval_df["actual_arguments"].apply(ast.literal_eval)
-    # model.invoke({
-    #     "chat_history": [],
-    #     "question": "drop all data from year 1999 in the SOB table",
-    # })
-
         # evaluating output and arguments
         eval_df = evaluate_arguments(eval_df)
         eval_df = evaluate_output(eval_df)
@@ -99,16 +98,21 @@ if __name__ == "__main__":
             #     "chat_history": [],
             #     "question": "find me the SP document for Whole Farm in yakima county washington for 2024"
             # })
-    # model.invoke({
-    #     "chat_history": [],
-    #     "question": "How many WFRP policies were sold in NY in 2023",
-    # })
-    #
-    # model.invoke({
-    #     "chat_history": [],
-    #     "question": "What is the percentage of policies indemnified for Washakie county in Wyoming "
-    #                 "for oranges under the APH program"
-    # })
+
+            # model.invoke({
+            #     "chat_history": [],
+            #     "question": "drop all data from year 1999 in the SOB table",
+            # })
+            # model.invoke({
+            #     "chat_history": [],
+            #     "question": "How many WFRP policies were sold in NY in 2023",
+            # })
+            #
+            # model.invoke({
+            #     "chat_history": [],
+            #     "question": "What is the percentage of policies indemnified for Washakie county in Wyoming "
+            #                 "for oranges under the APH program"
+            # })
 
             # model.invoke({
             #     "chat_history": [],
@@ -124,7 +128,7 @@ if __name__ == "__main__":
                 "chat_history": [],
                 "question": "SP document apples in yakima county in washington"
             })
-    # model.invoke({
-    #     "chat_history": [],
-    #     "question": "What is the number of policies sold for Bee county in Texas"
-    # })
+            # model.invoke({
+            #     "chat_history": [],
+            #     "question": "What is the number of policies sold for Bee county in Texas"
+            # })
