+++ main.py
@@ -14,8 +14,8 @@ from langserve import add_routes
 
 from pydantic.v1 import BaseModel
 
-from croptalk.model_llm import model
-# from croptalk.model_openai_functions import model
+#from croptalk.model_llm import model
+from croptalk.model_openai_functions import model
 
 
 set_debug(True)
